% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{de}
\alias{de}
\title{Node descendant identifier}
\usage{
de(x, tree)
}
\arguments{
\item{x}{character indicating node}

\item{tree}{tree structure generated from rcDT()}
}
\value{
Returns all the descendant nodes from the node `x` as a character string
}
\description{
Identifies all nodes in a given tree structure
which are descendants of the input node.
}
