% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Example data for rcITR package demonstration}
\usage{
generateData(n = 1000)
}
\arguments{
\item{n}{numeric. Number of example samples to be generated. Defaults to 1000.}
}
\value{
data.frame with 10 covariates (x1-x10), efficacy (y), risk (r), 
        treatment indicator (trt), and propensity score (prtx).
}
\description{
Generates data for demonstration of rcDT and rcRF model construction.
}
\examples{
dat <- generateData()
}
