% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrtest.R
\name{itrtest}
\alias{itrtest}
\title{Value function used for treatment assignment value calculation.}
\usage{
itrtest(dat, z, n0, aug)
}
\arguments{
\item{dat}{dataset being assessed}

\item{z}{new (alternative) treatment assignment in the splitting procedure.}

\item{n0}{minimum number of observations needed to make a split.}

\item{aug}{logical indicator for AIPWE estimation (TRUE) or IPWE estimation (FALSE).}
}
\value{
ITR value from the new treatment assignments
}
\description{
This function is used inside the tree growing functions.
}
