% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.set.R
\name{power.set}
\alias{power.set}
\title{Categorical predictor subsetting in the tree growing function.}
\usage{
power.set(x)
}
\arguments{
\item{x}{a categorical predictor.}
}
\value{
List of unique permutations of predictor `x`
}
\description{
Finds unique combinations of categorical, non-ordinal variables.
}
