% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Prunes an rcDT model to determine set of optimally pruned subtrees}
\usage{
prune(tre, risk.control = TRUE, risk.threshold = NA, lambda = lambda,
  a = 0, test = NULL, AIPWE = FALSE, n0 = 5, ctgs = NULL)
}
\arguments{
\item{tre}{sets the tree to be pruned}

\item{a}{sets the value of the splitting penalty}

\item{test}{the testing data to be used.  Defaults to NULL.}

\item{AIPWE}{indicator for AIPWE estimation.}

\item{n0}{minimum number of observations allowed in a treatment group. Defaults to 5.}

\item{ctgs}{columns of categorical variables.}

\item{train}{the training data used to create the tree}
}
\value{
summary of pruned branches and the associated value of the tree after pruning.

\item{result}{contains columns: `node.rm` which is the weakest link at each
iteration of the pruning algorithm; `size.tree` and `n.tmnl` give number of total nodes and
terminal nodes in each subtree; `alpha` is the weakest link scoring criteria; `V` and `V.test`
are the overall value of the tree for the training and tesing samples; `V.a` and `Va.test`
give the penalized value for the training and testing samples.}

\item{subtrees}{list of optimally pruned subtrees of `tre`}
}
\description{
The `prune` function determines the "weakest link" criteria for a given tree
in order to evaluate the order in which branches are pruned.
}
