% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcDT.select.R
\name{rcDT.select}
\alias{rcDT.select}
\title{Cross Validation for Optimal rcDT Model Selection for Given Penalty (lambda)}
\usage{
rcDT.select(dat, split.var, N0 = 20, n0 = 5, efficacy = "y", risk = "r",
  col.trt = "trt", col.prtx = "prtx", lambda = 0, risk.control = TRUE,
  risk.threshold = NA, nfolds = 10, AIPWE = FALSE, sort = TRUE,
  ctg = NA, max.depth = 15, stabilize.type = c("linear", "rf"),
  stabilize = TRUE, use.other.nodes = TRUE, use.bootstrap = FALSE,
  extremeRandomized = FALSE)
}
\arguments{
\item{dat}{data.frame. Data used to construct rcDT model.
Must contain efficacy variable (y), 
risk variable (r), 
binary treatment indicator coded as 0 / 1 (trt), 
propensity score (prtx),
candidate splitting covariates.}

\item{split.var}{numeric vector. Columns of spliting variables.}

\item{N0}{numeric specifying minimum number of observations required to call a node terminal. Defaults to 20.}

\item{n0}{numeric specifying minimum number of treatment/control observations needed in a split to declare a node terminal. Defaults to 5.}

\item{efficacy}{char. Efficacy outcome column. Defaults to 'y'.}

\item{risk}{char. Risk outcome column. Defaults to 'r'.}

\item{col.trt}{char. Treatment indicator column name. Should be of form 0/1 or -1/+1.}

\item{col.prtx}{char. Propensity score column name.}

\item{lambda}{numeric. Penalty parameter for risk scores. Defaults to 0, i.e. no constraint.}

\item{risk.control}{logical. Should risk be controlled? Defaults to TRUE.}

\item{risk.threshold}{numeric. Desired level of risk control.}

\item{AIPWE}{logical. Should AIPWE (TRUE) or IPWE (FALSE) be used. Not available yet.}

\item{sort}{internal use.}

\item{ctg}{numeric vector corresponding to the categorical input columns.  Defaults to NULL.  Not available yet.}

\item{max.depth}{numeric specifying maximum depth of the tree. Defaults to 15 levels.}

\item{stabilize.type}{character specifying method used for estimating residuals. Current options are 'linear' for linear model (default) and 'rf' for random forest.}

\item{stabilize}{logical indicating if efficacy should be modeled using residuals. Defaults to TRUE.}

\item{use.other.nodes}{logical. Should global estimator of objective function be used. Defaults to TRUE.}

\item{use.bootstrap}{logical. Should a bootstrap resampling be done? Defaults to FALSE.}

\item{extremeRandomized}{logical. Experimental for randomly selecting cutpoints in a random forest model. Defaults to FALSE and users should change this at their own peril. #' @return A summary of the cross validation including optimal penalty parameter and the optimal model.}

\item{test}{data.frame of testing observations. Should be formatted the same as 'data'.}

\item{mtry}{numeric specifying the number of randomly selected splitting variables to be included. Defaults to number of splitting variables.}
}
\value{
\item{best.tree.size}{optimal rcDT model based on size}

\item{best.tree.alpha}{optimal rcDT model based on alpha parameter selection}

\item{best.alpha}{optimal lambda parameter selected from the cross validation procedure}

\item{full.tree}{unpruned tree}

\item{pruned.tree}{output from pruning of `full.tree`}

\item{data}{input data}

\item{details}{summary of model performance}

\item{subtrees}{sequence of optimally pruned subtrees of `full.tree`}

\item{in.train}{training samples from splits}

\item{in.test}{testing samples from splits}
}
\description{
Performs k-fold cross validation for rcDT model to 
select the best subtree from the set of optimally pruned subtree generated from 
`prune` function.
}
\examples{

# Grow large tree
set.seed(1)
dat <- generateData()
fit <- rcDT.select(dat, split.var = 1:10, nfolds = 5, lambda = 1,
                   risk.control = TRUE, risk.threshold = 2.75)

}
