% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.partition.ITR.R
\name{risk.partition.ITR}
\alias{risk.partition.ITR}
\title{Generates partition summary based on itr value.}
\usage{
risk.partition.ITR(dat, split.var, test = NULL, risk.threshold = NA,
  min.ndsz = 20, n0 = 5, lambda = 0, name = "0", ctg = ctg,
  max.depth = 15, mtry = length(split.var), dat.rest = NULL,
  max.score = NULL, AIPWE = AIPWE, use.other.nodes = TRUE,
  extremeRandomized = FALSE)
}
\arguments{
\item{dat}{data.frame. Data used to identify split.}

\item{split.var}{numeric vector. Columns of spliting variables.}

\item{test}{data.frame of testing observations. Should be formatted the same as 'data'.}

\item{risk.threshold}{numeric. Desired level of risk control.}

\item{min.ndsz}{numeric specifying minimum number of observations required to call a node terminal. Defaults to 20.}

\item{n0}{numeric specifying minimum number of treatment/control observations needed in a split to declare a node terminal. Defaults to 5.}

\item{lambda}{numeric. Penalty parameter for risk scores. Defaults to 0, i.e. no constraint.}

\item{name}{char. Name of internal node, used for ordering splits.}

\item{ctg}{numeric vector corresponding to the categorical input columns.  Defaults to NULL.  Not available yet.}

\item{max.depth}{numeric specifying maximum depth of the tree. Defaults to 15 levels.}

\item{mtry}{numeric specifying the number of randomly selected splitting variables to be included. Defaults to number of splitting variables.}

\item{dat.rest}{dataframe. Data outside current splitting node.}

\item{max.score}{numeric. Current score for the tree.}

\item{AIPWE}{logical. Should AIPWE (TRUE) or IPWE (FALSE) be used. Not available yet.}

\item{use.other.nodes}{logical. Should global estimator of objective function be used. Defaults to TRUE.}

\item{extremeRandomized}{logical. Experimental for randomly selecting cutpoints in a random forest model. Defaults to #' @return summary of the best split for a given data frame.}
}
\description{
Determines optimal split for a given node.
}
