% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.itrtest.R
\name{s.itrtest}
\alias{s.itrtest}
\title{Estimates value for survival data (currently experimental)}
\usage{
s.itrtest(dat, z, n0, aug)
}
\arguments{
\item{dat}{dataset being assessed}

\item{z}{new (alternative) treatment assignment in the splitting procedure.}

\item{n0}{minimum number of observations needed to make a split.}

\item{aug}{logical indicator for AIPWE estimation (TRUE) or IPWE estimation (FALSE).}
}
\value{
ITR value from the new treatment assignments
}
\description{
This function is used inside the rcDT construction functions.
}
\examples{
dat <- gdataM(n=1000, depth=2, beta1=3, beta2=1)
# Assign every other patient to treatment
z <- rep(c(0,1), nrow(dat)/2)
# IPWE value
itrtest(dat, z, 5, FALSE)
# AIPWE value
itrtest(dat, z, 5, TRUE)
}
