% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.down.R
\name{send.down}
\alias{send.down}
\title{Sends testing data down a tree to obtain terminal node assignments}
\usage{
send.down(dat.new, tre, char.var = 1000, ctgs = NULL)
}
\arguments{
\item{dat.new}{data to be run down the tree.  Required input.}

\item{tre}{tree object from grow.ITR().  Required input.}

\item{char.var}{internal variable.}

\item{ctgs}{categorical variables, entered as columns in `dat.new`}
}
\value{
\item{data}{input data with extra column of node assignments}

\item{tree}{input tree with extra column for number of observations in each node}
}
\description{
Sends dat.new down tree 'tre' to obtain node assignment.
}
