% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.down.VI.ITR.Efficacy.R
\name{send.down.VI.ITR.Efficacy}
\alias{send.down.VI.ITR.Efficacy}
\title{Sends testing data and permuted testing data down a bootstrap tree from an rcRF object for efficacy scores.}
\usage{
send.down.VI.ITR.Efficacy(dat.new, tre, col.y, col.r, col.trt, col.prtx,
  risk.threshold, lambda, ctg = NULL, n0 = n0, N0 = N0, revise.tree = T,
  depth = 1, AIPWE = AIPWE)
}
\arguments{
\item{dat.new}{the new data set being sent down the tree. Required input.}

\item{tre}{constructed tree.}

\item{col.y}{the response variable. Required input.}

\item{col.trt}{the treatment indicator.  Must be binary. Required input.}

\item{col.prtx}{the probability of being assigned to treatment group. Required input.}

\item{ctg}{identifies the categorical input columns.  Defaults to NA.  Not available yet.}

\item{n0}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 5.}

\item{N0}{minimum number of observations needed in a node to stop splitting.}

\item{revise.tree}{internal variable.}

\item{depth}{internval variable}

\item{AIPWE}{indicator for AIPWE estimation.}
}
\value{
\item{tre0}{input tree with score from original tree and score from permutation from each variable used in the tree}

\item{score}{score from the permuted data}
}
\description{
Sends dat.new down a tree inside a random forest. 
Calcuates value and permuted value for variable importance measures
}
