% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.down.VI.ITR.R
\name{send.down.VI.ITR}
\alias{send.down.VI.ITR}
\title{Total importance via oob permutation}
\usage{
send.down.VI.ITR(dat.new, tre, col.y, col.r, col.trt, col.prtx, risk.threshold,
  lambda, ctg = NULL, n0 = n0, N0 = N0, revise.tree = T, depth = 1,
  AIPWE = AIPWE)
}
\arguments{
\item{dat.new}{the new data set being sent down the tree. Required input.}

\item{tre}{tree from rcDT function.}

\item{col.y}{the efficacy variable. Required input.}

\item{col.r}{the risk variable. Required input.}

\item{col.trt}{the treatment indicator.  Must be binary. Required input.}

\item{col.prtx}{the probability of being assigned to treatment group. Required input.}

\item{lambda}{numeric value for risk penalty.}

\item{ctg}{identifies the categorical input columns.  Defaults to NA.  Not available yet.}

\item{n0}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 5.}

\item{N0}{minimum number of observations needed in a node to stop splitting.}

\item{revise.tree}{internal variable.}

\item{depth}{internval variable}

\item{AIPWE}{indicator for AIPWE estimation.}
}
\value{
List of importance items

\item{tre0}{input tree with score from original tree and score from permutation from each variable used in the tree}

\item{score}{score from the permuted data}
}
\description{
Sends dat.new down an rcDT structure inside an rcRF
Calcuates value and permuted value for variable importance measures
}
