% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable.Importance.ITR.R
\name{Variable.Importance.ITR}
\alias{Variable.Importance.ITR}
\title{Calcuates variable importance measures for an rcRF object.}
\usage{
Variable.Importance.ITR(RF.fit, n0 = 5, N0 = 20, sort = TRUE,
  details = FALSE, depth = 1, AIPWE = FALSE)
}
\arguments{
\item{RF.fit}{forest object from rcRF(). Required input.}

\item{n0}{minimum number of treatment/control observations needed in a split to call a node terminal. Defaults to 2.}

\item{N0}{minimum number of observations needed in a split to call a node terminal. Defaults to 20.}

\item{sort}{sort the variable importance measure? Defaults to TRUE.}

\item{details}{print details of each tree as the function progresses. Defaults to FALSE.}

\item{depth}{internal variable.}

\item{AIPWE}{indicator for AIPWE estimation.}
}
\value{
Returns list of total (VI), efficacy (VI.Efficacy), and risk (VI.Risk) 
        ordered variable importance measure calculated for each splitting variable.
}
\description{
This function accepts a forest object from the `rcRF()` function and estimates the importance of each
predictor. This is accomplished by considering each tree in the forest, obtaining the out-of-bag value for each predictor in that tree, 
obtaining the permuted out-of-bag value for each predictor in the tree, and comparing the values. A larger discrepancy between 
the original value and permuted value indicates the predictor is more important in predicting treatment. The function
returns the variables in order of importance along with the importance measure, scaled to be out of 1.
}
\examples{
set.seed(123)
dat <- generateData(n = 1000)
# Build a forest with 100 trees
fit <- rcRF(dat = dat, split.var = 1:10, ntree = 200,
            risk.control = TRUE, risk.threshold = 2.75, 
            lambda = 1)
VI <- Variable.Importance.ITR(fit)
}
