% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{de}
\alias{de}
\title{Finds all decendants of a particular node in a tree.}
\usage{
de(x, tree)
}
\arguments{
\item{x}{node}

\item{tree}{interaction tree}
}
\value{
Returns all the descendant nodes from the node `x`
}
\description{
This function identifies all nodes in a given tree structure
which are descendants of the input node.
}
