% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Optimal sequence of subtrees of rcDT model}
\usage{
prune(tre, risk.control = TRUE, risk.threshold = NA, lambda = lambda,
  a = 0, test = NULL, AIPWE = FALSE, n0 = 5, ctgs = NULL)
}
\arguments{
\item{tre}{sets the tree to be pruned}

\item{risk.control}{logical indicating if risk controlled model is under consideration}

\item{risk.threshold}{numeric indicating the value of the risk control. Defaults to NA.}

\item{a}{numeric value of the splitting penalty. Defaults to zero.}

\item{test}{data.frame of testing data. Defaults to NULL.}

\item{AIPWE}{indicator for AIPWE estimation.}

\item{n0}{minimum number of observations allowed in a treatment group. Defaults to 5.}

\item{ctgs}{columns of categorical variables.}
}
\value{
summary of sequence of subtrees

\item{result}{contains columns: `node.rm` which is the weakest link at each
iteration of the pruning algorithm; `size.tree` and `n.tmnl` give number of total nodes and
terminal nodes in each subtree; `alpha` is the weakest link scoring criteria; `V` and `V.test`
are the overall value of the tree for the training and tesing samples; `V.a` and `Va.test`
give the penalized value for the training and testing samples. `Benefit` and `Risk` scores
are also reported.}

\item{subtrees}{list of optimally pruned subtrees of `tre`}
}
\description{
Determines the sequence of optimally pruned subtrees for 
             an rcDT model.
}
